dashboard: First shift towards Qt's Model/View framework
Some logic of loading data etc. will be moved to the model, making it
reusable and flexible.

The Model/View structure makes it possible to use a table, with proper
alignment - once done it'll look better than the messy custom widgets we
had before.

An alternative could be using a simple QTableWidget with custom widgets
for text formatting, but I think using delegates is nicer.