dashboard: implement "stretchFirstSection"-like functionality
Qt provides stretchLastSection property in the QHeaderView, but nothing
like stretchFirstSection. This implementation kinda implements it, with
a small twist: when resizing the first column manually, that'll become
the minimal size for auto-resizing. When the column is automatically
resized, it won't become smaller than the interactively chosen size.

The main difference is that the column is not stretched to a size such
that the viewport is always filled. Instead it's stretched along with
the viewport, so there may be some empty space to the right of the table
even with this stretching.