dashboard: Compute ahead/behind colors for better contrast
A fairly simple formula computes the lightness to use for the text
colors, based on the background colors indicated by QPalette().
Essentially it's as follows:

* Gradually increase lightness in the range [.50, .75] if background
lightness <= .50
* Gradually decrease lightness in range [.50, .25] if background
lightness > .50

The range is limited to .25-.75 to keep some color and prevent full
black or white colors. Tested and works well enough for basic dark or
light color schemes; it only becomes somewhat hard to read when using
one of the basic colors (red/yellow/green) as background color.